Ex03- Advaced Math version 1

+++ src/com/mta/javacourse/Gal_NasdaqServlet.java
@@ -28,7 +28,7 @@ public class Gal_NasdaqServlet extends HttpServlet {
 			
 		opposite=30;
 		hypotenuse=50;
-		int result2= opposite / hypotenuse ;
+		double result2= opposite / hypotenuse ;
 		
 		String line2= new String ("<h1> Calculation 2: Length of opposite where angle B is 30 degrees and Hypotenuse length is 50 cm is: "+opposite+" / "+hypotenuse+" = "+result2+"  cm </h1>");	
 		resp.getWriter().println(line2);
